+++ inertia_svelte.rb
@@ -136,6 +136,7 @@ after_bundle do
   apply "src/inertia_shared/gems_anyway_config.rb"
   apply "src/shared/gems_pagy.rb"
   apply "src/shared/gems_active_interaction.rb"
+  apply "src/inertia_shared/gems_oj_serializers.rb"
 
   apply "src/shared/gems_rspec.rb"
   apply "src/inertia_shared/rspec_inertia.rb"